refactor: foxy deposit split
