+++ src/features/defi/providers/foxy/components/FoxyManager/Deposit/DepositCommon.ts
@@ -1,0 +1,8 @@
+export enum DepositPath {
+  Deposit = '/',
+  Approve = '/approve',
+  ApproveSettings = '/approve/settings',
+  Confirm = '/confirm',
+  ConfirmSettings = '/confirm/settings',
+  Status = '/status',
+}
